==> data_processing/4_preprocessed_data.ipynb <==
%config Completer.use_jedi = False



import json
import pandas as pd
import os
from tqdm import tqdm
import warnings
warnings.filterwarnings(action='ignore')
path = 'gameinfo/'
file_list = os.listdir(path)
file_list_py = [file for file in file_list if file.endswith('.json')]

file_list_py

all_champ_data = pd.DataFrame()
for fpath in tqdm(file_list_py):
    with open ("gameinfo/"+fpath, "r", encoding="utf-8-sig") as json_file:
        jdata = json.load(json_file)
    jArray = jdata.get("info").get("participants")
    one_game_data = pd.DataFrame()
    for i in jArray:
        champ_data = pd.DataFrame()
        Position = i.get("individualPosition")
        Champ_Name = i.get("championName")
        Champ_ID = i.get("championId")
        item0 = i.get("item0")
        item1 = i.get("item1")
        item2 = i.get("item2")
        item3 = i.get("item3")
        item4 = i.get("item4")
        item5 = i.get("item5")
        item6 = i.get("item6")
        win = i.get("win")
        champ_data["Position"] = [Position]
        champ_data["Champ_Name"] = [Champ_Name]
        champ_data["Champ_ID"] = [Champ_ID]
        champ_data["item0"] = [item0]
        champ_data["item1"] = [item1]
        champ_data["item2"] = [item2]
        champ_data["item3"] = [item3]
        champ_data["item4"] = [item4]
        champ_data["item5"] = [item5]
        champ_data["item6"] = [item6]
        champ_data["win"] = [win]
        
        one_game_data = one_game_data.append(champ_data)

    temp_df = one_game_data[['Champ_ID']].iloc[5:10]
    temp_df = temp_df.append(one_game_data[['Champ_ID']].iloc[0:5])
    one_game_data['vs']= temp_df
    all_champ_data = all_champ_data.append(one_game_data)

all_champ_data = all_champ_data.reset_index()

all_champ_data = all_champ_data.drop('index',axis = 1)

all_champ_data

all_champ_data.to_csv("game_data.csv")
